feat(tasks): Add Process Locks from MSR
