+++ package.json
@@ -13,7 +13,7 @@
 		"create-task:gm": "hardhat run ./scripts/create-gm-task.ts --network ethereum",
 		"create-task:spell-staking": "killall -9 anvil; hardhat run ./scripts/create-spell-staking-task.ts --network anvil",
 		"create-task:magiclvl": "hardhat run ./scripts/create-magiclvl-task.ts --network anvil",
-		"create-task:mim-saving-rate": "hardhat run ./scripts/create-mim-saving-rate-task.ts --network arbitrum",
+		"create-task:process-locks": "hardhat run ./scripts/create-process-locks-task.ts --network arbitrum",
 		"create-task:spell-swapper": "hardhat run ./scripts/create-spell-swapper-task.ts --network ethereum",
 		"create-task:negative-interests": "hardhat run ./scripts/create-negative-interests-task.ts --network anvil",
 		"create-task:reward-distributor": "hardhat run ./scripts/create-reward-distributor-task.ts --network arbitrum",
@@ -35,7 +35,8 @@
 		"test:gm:gmlink": "cp web3-functions/gm/userArgs.gmlink.json web3-functions/gm/userArgs.json && hardhat w3f-run gm --logs --network arbitrum",
 		"test:gm:gmsol": "cp web3-functions/gm/userArgs.gmsol.json web3-functions/gm/userArgs.json && hardhat w3f-run gm --logs --network arbitrum",
 		"test:magiclvl": "hardhat w3f-run magiclvl --logs --network bsc",
-		"test:mim-saving-rate": "hardhat w3f-run mim-saving-rate --logs --network arbitrum",
+		"test:process-locks-arb-msr": "cp web3-functions/process-locks/userArgs.arb-msr.json web3-functions/process-locks/userArgs.json && hardhat w3f-run process-locks --logs --network arbitrum",
+		"test:process-locks-blast-founders": "cp web3-functions/process-locks/userArgs.blast-founders.json web3-functions/process-locks/userArgs.json && hardhat w3f-run process-locks --logs --network arbitrum",
 		"test:spell-swapper": "hardhat w3f-run spell-swapper --logs --network ethereum",
 		"test:negative-interests": "cp web3-functions/negative-interests/userArgs.default.json web3-functions/negative-interests/userArgs.json;  hardhat w3f-run negative-interests --logs --network ethereum; rm web3-functions/negative-interests/userArgs.json",
 		"test:reward-distributor": "hardhat w3f-run reward-distributor --logs --network arbitrum",
