+++ web3-functions/process-locks/index.ts
