+++ web3-functions/process-locks/schema.json
