+++ web3-functions/process-locks/userArgs.blast-founders.json
@@ -1,0 +1,5 @@
+{
+	"execAddress": "0xF1eD28fa139f2dF5CF3ed140aa9F803C79554519",
+	"subgraphUrl": "https://api.studio.thegraph.com/query/59137/founder-locks/version/latest",
+	"lockLimitPerExecution": 100
+}
